Split Assistant services
